Use the supported time profiling method, i.e. real time, on Cygwin by default
Branches: main, 11.01

Use the supported time profiling method, i.e. real time, on Cygwin by default
rather than attempting to use an unsupported method.

Avoid a C compiler warning on Cygwin.

runtime/mercury_wrapper.c:
	Make real time the default time profiling method on Cygwin since
	the other methods are not supported on that platform.
	(According to the Cygwin implementation notes on only ITIMER_REAL
	is supported with setitimer.)

doc/user_guide.texi:
	Mention that -Tr is the default time method for time
	profiling on Cygwin.  (The existing text referred to
	Windows, but that's not accurate - time profiling doesn't
	work at all with MinGW.)

	Unrelated change: Erlang is also supported as a foreign
	language.

compiler/c_util.m:
	Avoid a warning from the C compiler on Cygwin.

+++ compiler/c_util.m
@@ -280,7 +280,7 @@ can_print_directly(_, no, !IO).
         can_print_directly = MR_TRUE;
 
         for (s = Str; *s != '\\0'; s++) {
-            if (! (isalnum(*s) || *s == '_' || *s == '/' || *s == '.')) {
+            if (! (isalnum((int)*s) || *s == '_' || *s == '/' || *s == '.')) {
                 can_print_directly = MR_FALSE;
                 break;
             }
